refactor: better way to get movies with query params
The MoviesViewModel now has a new function -
getMoviesWithQueryParams, which takes in not a single queryParams
object, but instead, all the parameters as distinct objects.
This way you only need to provide the parameters you care about,
while the omitted ones are provided by the viewModel, which were saved
in the queryParams field with previous queries.