fix: incorrect movie info binding
Swapped the implementation of fetching movie info from db and network.
Instead of observing LiveData objects, the data is requested directly,
by launching coroutines. If an error occures, a toast message is shown.
Fixed a bug when there are two duplicates of trailers due to additional
English language parameter hardcoded in the network query.
Place a "null" instead temporarily.